+++ app/views/assignment_repos/show.html.erb
@@ -1,4 +1,4 @@
-<% view = AssignmentRepo::ShowView.new(assignment_repo: @assignment_repo) %>
+<% view = AssignmentRepoView::ShowView.new(assignment_repo: @assignment_repo) %>
 
 <div class='d-flex flex-row flex-justify-between flex-items-center assignment-repo-list-item'>
   <div class='d-flex'>
