+++ app/views/group_assignment_repos/show.html.erb
@@ -1,4 +1,4 @@
-<% view = GroupAssignmentRepo::ShowView.new(assignment_repo: @group_assignment_repo) %>
+<% view = GroupAssignmentRepoView::ShowView.new(assignment_repo: @group_assignment_repo) %>
 
 <div class='d-flex flex-row flex-justify-between flex-items-center assignment-repo-list-item'>
   <div class='d-flex flex-column flex-content-around'>
